pyproject.toml
[project]
name = "qml-analyzinggalaxies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "astropy>=7.2.0",
    "datasets>=4.8.4",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "polars>=1.39.3",
    "pyarrow>=23.0.1",
    "requests>=2.33.0",
    "scikit-learn>=1.8.0",
    "scipy-stubs>=1.17.1.3",
    "sdss-access>=3.0.10",
    "pennylane>=0.39.0",
    "torch>=2.6.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }